Move `relay/http/server` into `relay/server/http_server`
